Fixing a lot of bugs
